Fixed a warning in `flags.hpp`.
Removes an implicit cast from `size_t` to `int`.

Review: https://reviews.apache.org/r/52205/

+++ 3rdparty/stout/include/stout/flags/flags.hpp
@@ -861,7 +861,7 @@ inline Try<Warnings> FlagsBase::load(
   // Update 'argc' and 'argv' if we successfully loaded the flags.
   if (!result.isError()) {
     CHECK_LE(args.size(), (size_t) *argc);
-    size_t i = 1; // Start at '1' to skip argv[0].
+    int i = 1; // Start at '1' to skip argv[0].
     foreach (char* arg, args) {
       (*argv)[i++] = arg;
     }
